# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Security Lab API - Gestión de Secretos y Escalamiento de Privilegios
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  mensaje: string;
  /** @nullable */
  campo?: string | null;
}

export interface RegisterInput {
  /** @minLength 2 */
  nombre: string;
  correo: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  correo: string;
  password: string;
}

export type UserProfileRol = typeof UserProfileRol[keyof typeof UserProfileRol];


export const UserProfileRol = {
  USUARIO: 'USUARIO',
  ADMIN: 'ADMIN',
} as const;

export interface UserProfile {
  id: number;
  nombre: string;
  correo: string;
  rol: UserProfileRol;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  usuario: UserProfile;
}

export interface ProfileUpdateUnsafe {
  nombre?: string;
  correo?: string;
  password?: string;
  rol?: string;
  estado?: string;
  permisos?: string;
}

export interface ProfileUpdateSafe {
  nombre?: string;
  correo?: string;
  password?: string;
}

export interface AdminStats {
  totalUsuarios: number;
  totalAdmins: number;
  registrosRecientes: number;
}