# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Security Lab API - Gestión de Secretos y Escalamiento de Privilegios
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  mensaje: string;
  /** @nullable */
  campo?: string | null;
}