# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Security Lab API - Gestión de Secretos y Escalamiento de Privilegios
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRol } from './userProfileRol';

export interface UserProfile {
  id: number;
  nombre: string;
  correo: string;
  rol: UserProfileRol;
  createdAt: Date;
}